device: Make get_guid() method private again
It turns out that GNOME Games doesn't need it in the end thanks to the
inclusion of the mapping accessors in ManetteDevice.

As it is easier to add something to a release than to remove it from
one, let's keep this method private until a user of the library requests
it.