device: Defer event emission
Defer the emission of input device events by putting them into idle
event sources.

This avoids the handling of an event to block the device's event polling
and processing until it's resolved. For example this prevents the press
of a button which triggers the apparition of a GtkDialog from blocking
the device until the dialog stops running, hence allowing to use the
device in that same dialog.