Use assertions instead of g_return*_if_fail in static functions
As g_return_if_fail is only for public API checking. If there need to be
some arguments checking in static functions, it is a programmer error
and so must assert.

This commit also removes type checking for self (i.e. the type defined
in this file), since those check should already be put for public API
functions.