event-mapping: Make hat to buttons mapping always emit button release
When a hat mapped to two buttons was quickly pressed, with only the
extreme values reported and not the neutral point, there wasn't any
button release event for the direction that was moved from, and so it
was still recorded as pressed.

This commits fixes it, by taking account the common event value and
sending a button release event when it's detected as going in the
opposite direction.