Introduce groundwork for multiple device types and backends
As nice as fully generic evdev code is, it's not enough for some gamepads,
particularly the built-in one in Steam Deck. So, for these gamepads we want
to do with e.g. SDL does and use HID instead.

However, right now evdev is very tightly integrated into ManetteDevice, so
first we need to split it out. For that, introduce a delegate interface
called ManetteBackend.

Right now ManetteBackend only has a single implementation:
ManetteEvdevBackend. Later we'll have a HID one parallel to it.