Introduce HID backend and Steam Deck HID driver
Now that we have separate backends, implement a second one using HID.

HID devices are not generic and we have to write bespoke drivers for them.
I mean, they are in theory generic, but they really aren't in practice.

So, introduce a second interface ManetteHidDriver. ManetteHidBackend must
have a ManetteHidDriver instance to work, it just handles the common tasks.

Introduce a single ManetteHidDriver implementation, ManetteSteamDeckDriver.
It's heavily inspired by SDL and kernel hid-steam code, and handles all the
basic input for Steam Deck gamepad, while disabling lizard mode (only works
if Steam is not running, or when Highscore is launched from Steam itself).

Unlike SDL and hid-steam, it keeps touchpads fully working with lizard mode
off. We may want to change that later if we expose touchpad input directly,
but right now we don't, so let's have them enabled. Since libmanette is
usually used in desktop apps and not fullscreen games, pointer is very
useful to have.

There are still a few things we could expose that I didn't do here:

- Whether the user is touching each of the analog sticks
- Analog trigger values
- Coordinates and pressure for each touchpad
- Whether the user is touching and/or pressing each touchpad
- Accelerometer
- Gyro